Correct data type for Dataset['year']
- accidentally generated as float, changed to integer
- ran migrations again
- added seed data for datasets

+++ db/migrate/20170325012310_create_datasets.rb
@@ -1,8 +1,10 @@
+# frozen_string_literal: true
+
 class CreateDatasets < ActiveRecord::Migration[5.0]
   def change
     create_table :datasets do |t|
       t.text :name, null: false
-      t.numeric :year, null: false
+      t.integer :year, null: false
       t.text :resource, null: false
 
       t.timestamps
